Adding better styles for buttons
ItemsPanel was separated to its own file
