+++ challenge-001/src/App.jsx
@@ -1,5 +1,6 @@
 import { useState } from "react";
 import "./App.css";
+import ItemsPanel from "./ItemsPanel";
 
 function App() {
   const [count, setCount] = useState(0);
@@ -8,23 +9,18 @@ function App() {
     <>
       <h1>Items Columns</h1>
       <div className="card p-7">
-        <div className="container flex gap-2">
-          <ItemsPanel></ItemsPanel>
-          <ItemsPanel></ItemsPanel>
+        <div className="container flex gap-2 w-full">
+          <ItemsPanel />
+
+          <div className="buttons-container flex flex-col gap-3 justify-center">
+            <button> &#x27A1; </button>
+            <button> &#x2B05; </button>
+          </div>
+
+          <ItemsPanel />
         </div>
       </div>
     </>
-  );
-}
-
-function ItemsPanel() {
-  const items = ["first", "second", "third"];
-  return (
-    <div className="flex-1 bg-slate-600 h-40 w-14 rounded">
-      {items.map((item) => (
-        <div>{item}</div>
-      ))}
-    </div>
   );
 }
